Fix "save-always" option
The previous commit reverts PR #1452, "Deprecate save-always input",
and this commit fixes the "save-always" option.

The "save-always" input failed to force "post" script run because
`github.event.inputs` refers to workflow dispatch or workflow call
inputs, not the action's `inputs:` defined inside the `action.yml`, so
`github.event.inputs.save-always` was never true.

Turns out it's hard to get the job's status inside the action as well,
and there's no equivalent of the `success()` function.  The
workaround, passing `job.status` in as an input, was cribbed from
`actions/setup-java`, where the `job-status` input is defined in
`action.yml` [1], and passed through (in a loopy way) for retrieval
from a utility function [2].

[1]:  https://github.com/actions/setup-java/blob/f4f1212c88/action.yml#L62
[2]:  https://github.com/actions/setup-java/blob/f4f1212c88/src/util.ts#L86-L88